feat(subscription): enhance CreditGrantModal with billing cadence
