+++ src/components/organisms/index.ts
@@ -1,2 +1,11 @@
-export { PlanDetailsSection, SetupChargesSection, RecurringChargesForm } from './PlanForm';
 export { default as EmptyPage } from './EmptyPage';
+export {
+	AddonTable,
+	Preview,
+	PriceTable,
+	SubscriptionActionButton,
+	SubscriptionForm,
+	SubscriptionTable,
+	TimelinePreview,
+	UsageTable,
+} from './Subscription';
